Move getTimeLeft to dates utils
