+++ vue-app/src/views/ProjectList.vue
@@ -84,6 +84,8 @@ import { DateTime } from 'luxon'
 import { RoundInfo, getCurrentRound, TimeLeft } from '@/api/round'
 import { Project, getRecipientRegistryAddress, getProjects } from '@/api/projects'
 
+import { getTimeLeft } from '@/utils/dates'
+
 import ProjectListItem from '@/components/ProjectListItem.vue'
 import MatchingFundsModal from '@/components/MatchingFundsModal.vue'
 import CartWidget from '@/components/CartWidget.vue'
@@ -114,15 +116,6 @@ function shuffleArray(array: any[]) {
     const j = Math.floor(random(SHUFFLE_RANDOM_SEED, i) * (i + 1))
     ;[array[i], array[j]] = [array[j], array[i]]
   }
-}
-
-function timeLeft(date: DateTime): TimeLeft {
-  const now = DateTime.local()
-  if (now >= date) {
-    return { days: 0, hours: 0, minutes: 0 }
-  }
-  const { days, hours, minutes } = date.diff(now, ['days', 'hours', 'minutes'])
-  return { days, hours, minutes: Math.ceil(minutes) }
 }
 
 @Component({
@@ -201,11 +194,11 @@ export default class ProjectList extends Vue {
   }
 
   get contributionTimeLeft(): TimeLeft {
-    return timeLeft(this.$store.state.currentRound.signUpDeadline)
+    return getTimeLeft(this.$store.state.currentRound.signUpDeadline)
   }
 
   get reallocationTimeLeft(): TimeLeft {
-    return timeLeft(this.$store.state.currentRound.votingDeadline)
+    return getTimeLeft(this.$store.state.currentRound.votingDeadline)
   }
 
   addMatchingFunds(): void {
